requirements.txt
openai
chromadb
pandas
openpyxl
streamlit
python-dotenv